Move B1530A module to its own module KeysightB1530A.py
